requirements.txt
neo4j
python-dotenv
pyfastx
biopython
geopandas
rdkit
networkx
matplotlib
seaborn